// client/src/app/stores/customers/customer.types.ts
export enum CUSTOMERS_ACTION_TYPES {
    FETCH_ALL_CUSTOMERS_LOADING = "customers/FETCH_ALL_CUSTOMERS_LOADING",
    FETCH_ALL_CUSTOMERS_SUCCESS = "customers/FETCH_ALL_CUSTOMERS_SUCCESS",
    FETCH_ALL_CUSTOMERS_FAILED = "customers/FETCH_ALL_CUSTOMERS_FAILED",
    FETCH_CUSTOMER_BY_ID_LOADING = "customers/FETCH_CUSTOMER_BY_ID_LOADING",
    FETCH_CUSTOMER_BY_ID_SUCCESS = "customers/FETCH_CUSTOMER_BY_ID_SUCCESS",
    FETCH_CUSTOMER_BY_ID_FAILED = "customers/FETCH_CUSTOMER_BY_ID_FAILED",
    ADD_CUSTOMER_LOADING = "customers/ADD_CUSTOMER_LOADING",
    ADD_CUSTOMER_SUCCESS = "customers/ADD_CUSTOMER_SUCCESS",
    ADD_CUSTOMER_FAILED = "customers/ADD_CUSTOMER_FAILED",
    UPDATE_CUSTOMER_LOADING = "customers/UPDATE_CUSTOMER_LOADING",
    UPDATE_CUSTOMER_SUCCESS = "customers/UPDATE_CUSTOMER_SUCCESS",
    UPDATE_CUSTOMER_FAILED = "customers/UPDATE_CUSTOMER_FAILED",
    DELETE_CUSTOMER_LOADING = "customers/DELETE_CUSTOMER_LOADING",
    DELETE_CUSTOMER_SUCCESS = "customers/DELETE_CUSTOMER_SUCCESS",
    DELETE_CUSTOMER_FAILED = "customers/DELETE_CUSTOMER_FAILED",
  };
  